move flagged spots to parent component
